formal/requirements.txt
ply==3.11
torch>=1.6.0
torchvision
pytest==5.4.3
redis
pexpect
scipy
cython
pyzmq
tornado
pynvml
submitit
scikit-learn
psutil
typeguard
tensorboard
future
tqdm
py7zr
typer
Flask
numexpr
pandas
youtokentome
graphviz
python-magic
BeautifulSoup4
jupyter
treetable
mpld3
fire
mypy
nevergrad
pympler
GitPython
black  #==19.10b0
click==8.0.4
sympy
warcio
PyYAML
